feat(spring): main 댓글 삭제 기능 추가

+++ SpringBoot/src/main/java/com/project/machinlearning/Comment/CommentController.java
@@ -1,9 +1,12 @@
 package com.project.machinlearning.Comment;
 
 import com.project.machinlearning.Comment.DTO.CommentRequestDTO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.text.ParseException;
 
 /**
@@ -28,8 +31,13 @@ public class CommentController {
     }
 
     @DeleteMapping("/delete/{cid}") //댓글 삭제
-    public String deleteComment(@PathVariable("cid") Long cid){
-        return commentService.deleteComment(cid);
+    public @ResponseBody ResponseEntity deleteComment(@PathVariable("cid") Long cid, Principal principal){
+        if(!commentService.validateComment(cid, principal.getName())) {
+            return new ResponseEntity<String>("수정 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
+       commentService.deleteComment(cid);
+        return new ResponseEntity<Long>(cid, HttpStatus.OK);
     }
